Extract OO controllers for arbitary resource paths
As it turns out, settings was the only resource of this type. The change
creates an API-namespaced SettingsController, which defines `#show`,
overriding the BaseController's `#show` which was handling the routing
to find the previously defined `#show_settings`.
